Refactor comparison functions
